Refactor to less files
