+++ mix.exs
@@ -26,7 +26,7 @@ defmodule BeeBee.Mixfile do
   # Type "mix help compile.app" for more information
   def application do
     [
-      mod: {BeeBee.Application, []},
+      mod: {BeeBee, []},
       # Specify extra applications you'll use from Erlang/Elixir
       extra_applications: [:logger]
     ]
